PROD4POD-1315 Create new DataTypes component
