+++ features/polyExplorer/src/locales/de/stories/clusterDigitalGiants.json
@@ -4,7 +4,6 @@
     "intro.p.1": "Wenn Sie heutzutage ein digitales Leben führen, ist es fast unmöglich, nicht mit einigen der großen Technologieunternehmen und ihren Diensten in Kontakt zu kommen. Was diese Unternehmen gemeinsam haben, ist ihr enormer Appetit auf Daten und die Absicht, diese Daten - wahrscheinlich Ihre Daten - für Werbung oder anstehende Big-Data-Projekte zu nutzen und zu Geld zu machen.",
     "intro.p.2": "Das polyPedia-Team hat sich alles angesehen, was die unten aufgeführten Unternehmen über ihre Datenschutzbestimmungen sagen.",
     "data.types.p": "Einige Datentypen müssen für das, was wir von einem Boten erwarten, gemeinsam genutzt werden. Aber nicht alle Boten teilen die gleiche Menge an Datentypen. Die zusätzliche gemeinsame Nutzung zeigt, dass sie damit Geld verdienen.",
-    "data.types.legend": "data type",
     "data.types.tab.companies": "By Companies",
     "data.types.tab.shares": "By Shares",
     "data.types.tab.types": "By Types",
