+++ features/polyExplorer/src/locales/de/stories/common.json
@@ -6,6 +6,7 @@
     "purposes.barChart.legendTitle": "Number of mentions",
     "purposes.barChart.max": "Maximum",
     "purposes.barChart.average": "Durchschnitt",
+    "data.types.legend": "data type",
     "section.companies": "Unternehmen",
     "section.dataRegions": "Daten Regionen",
     "section.explore.further": "Erkunde Weiter",
