+++ features/polyExplorer/src/locales/en/stories/clusterMessengerStory.json
@@ -25,7 +25,6 @@
     "details.monthly.active.users": "Monthly active users (MAU in millions)",
     "details.monthly.active.users.21": "Monthly active users 2020/2021 in millions",
     "data.types.title": "Data Types",
-    "data.types.legend": "data type",
     "data.types.tab.messenger": "By Messenger",
     "data.types.text.messenger": "Some data types need to be shared for what we expect from a messenger. But not all messengers share the same amount of data types. The extra sharing indicates that they make money with it.",
     "data.types.tab.shares": "By Shares",
