+++ features/polyExplorer/src/locales/en/stories/common.json
@@ -6,6 +6,7 @@
     "purposes.barChart.legendTitle": "Number of mentions",
     "purposes.barChart.max": "Maximum",
     "purposes.barChart.average": "Average",
+    "data.types.legend": "data type",
     "section.companies": "Companies",
     "section.dataRegions": "Data Regions",
     "section.explore.further": "Explore Further",
